requirements.txt
flask
flask-cors
python-dotenv
requests
PyPDF2
python-docx
flask-swagger
flasgger
PyYAML
